+++ src/main/java/com/company/project/web/ThCoinTransferItemController.java
@@ -15,7 +15,7 @@ import javax.annotation.Resource;
 import java.util.List;
 
 /**
-* Created by CodeGenerator on 2018/10/02.
+* Created by CodeGenerator on 2018/10/03.
 */
 @RestController
 @RequestMapping("/th/coin/transfer/item")
@@ -30,7 +30,7 @@ public class ThCoinTransferItemController {
     }
 
     @PostMapping("/delete")
-    public Result delete(@RequestParam Integer id) {
+    public Result delete(@RequestParam Long id) {
         thCoinTransferItemService.deleteById(id);
         return ResultGenerator.genSuccessResult();
     }
@@ -42,7 +42,7 @@ public class ThCoinTransferItemController {
     }
 
     @PostMapping("/detail")
-    public Result detail(@RequestParam Integer id) {
+    public Result detail(@RequestParam Long id) {
         ThCoinTransferItem thCoinTransferItem = thCoinTransferItemService.findById(id);
         return ResultGenerator.genSuccessResult(thCoinTransferItem);
     }
